healthsync/llm_api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-generativeai==0.4.0
pymongo==4.6.1
python-dotenv==1.0.0
pydantic==2.6.4